requirements.txt
fastmcp
uvicorn
starlette
httpx 